fix: el moverFicha envía enteros en vez de float

+++ app/src/main/java/upc/edu/pe/desaplg/connection/JsonHelper.java
@@ -104,7 +104,7 @@ public class JsonHelper {
         }
     }
 
-    public static JSONObject moverFicha(final float desX, final float desY){
+    public static JSONObject moverFicha(final int desX, final int desY){
 
         try {
             final JSONObject desplazamiento = new JSONObject(){
